requirements.txt
Flask==3.0.3
numpy==1.24.0
Pillow==10.3.0
pytesseract==0.3.10
scikit-image==0.23.2
ultralytics==8.2.31
matplotlib==3.9.0
symspellpy==6.7.7
gunicorn
google-cloud-storage
tensorflow
keras
pandas
scikit-learn
seaborn
mysql-connector-python
sqlalchemy
pymysql
python-dotenv
gunicorn